requirements.txt
flask==3.0.0
requests==2.31.0
opentelemetry-distro==0.45b0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-instrumentation-flask==0.45b0

